kiss.el: add kiss--silent-shell-command & try debugging install
when installing a package that puts files into the alternatives directory,
there is a bug where kiss will create a bunch of files in the current
`default-directory` that will then (later on) cause problems w/ the install.

The files that it creates are empty, and the names originate from the name that
they would have in the choices directory, ie 'usr', 'bin', 'share', etc.

This hasn't been a *huge* problem yet, but it is extremely annoying, and,
if we were to assert that every command in the install needs to be successful
then this would cause problems.